requirements.txt
gspread==6.0.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
